moved public header to include/...
without this change headers are not included with "QnTools/HEADER.hpp" but only as "HEADER.hpp"
this change was necessary because of root dictionary generation logic. relative path of files has to be the same during generation and use.
moved public header to include/...

without this change headers are not included with "QnTools/HEADER.hpp" but only as "HEADER.hpp"
this change was necessary because of root dictionary generation logic. relative path of files has to be the same during generation and use.
cleans up cmake 
root targets include the interface_include_directory property already